campo salary corregido en la migración 2024_03_06_124211_add_location_and_salary_to_jobs_table y ajustes en la funcionalidad de crear jobs de acuerdo a esta corrección

+++ database/migrations/2024_03_06_124211_add_location_and_salary_to_jobs_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::table('jobs', function (Blueprint $table) {
             $table->string('location')->nullable();
-            $table->decimal('salary', 8, 2)->nullable();
+            $table->integer('salary')->nullable();
         });
     }
